+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bug Fixes
 
+* Restored WFieldWarningIndicator and WFieldErrorIndicator to WField for backwards compatibility.
+
 ## Release 1.4.15
 
 ### Enhancements